refactor: mudança de local dos dtos, handlers e factories

+++ TimeProject.Application/UseCases/Auths/LoginGithubUseCase.cs
@@ -1,14 +1,15 @@
 ﻿using Octokit;
+using TimeProject.Application.Dtos.Auths;
+using TimeProject.Application.Dtos.Users;
+using TimeProject.Application.Interfaces.Handlers;
+using TimeProject.Application.Interfaces.Shared;
 using TimeProject.Application.Interfaces.UseCases.CustomLogs;
 using TimeProject.Application.Interfaces.UseCases.Logins;
 using TimeProject.Application.Interfaces.UseCases.Users;
+using TimeProject.Application.Shared;
 using TimeProject.Domain.Entities;
-using TimeProject.Domain.Shared;
 using TimeProject.Infrastructure.Errors;
 using TimeProject.Infrastructure.Interfaces;
-using TimeProject.Infrastructure.ObjectValues;
-using TimeProject.Infrastructure.ObjectValues.Auths;
-using TimeProject.Infrastructure.ObjectValues.Users;
 
 namespace TimeProject.Application.UseCases.Auths;
 
@@ -37,7 +38,7 @@ public class LoginGithubUseCase(
 
             if (getUserByPIdResult is { Data: not null })
             {
-                ac.UserId = (int)getUserByPIdResult.Data.UserId!;
+                ac.UserId = getUserByPIdResult.Data.UserId!;
                 createUserAccessLogUseCase.Handle(ac);
                 return result.SetData(jwtHandler.Generate(getUserByPIdResult.Data));
             }
